+++ core/src/main/java/org/apache/struts/validator/ValidatorForm.java
@@ -103,11 +103,22 @@ public class ValidatorForm extends ActionForm implements Serializable {
      */
     public ActionErrors validate(ActionMapping mapping,
         HttpServletRequest request) {
-        ServletContext application = getServlet().getServletContext();
-        ActionErrors errors = new ActionErrors();
 
+        ActionErrors errors = new ActionErrors();
         String validationKey = getValidationKey(mapping, request);
 
+        ServletContext application;
+        try {
+            application = getServlet().getServletContext();
+        } catch (NullPointerException e) {
+            IllegalStateException e2 = new IllegalStateException(
+                    "Missing ActionServlet instance for bean '" +
+                    mapping.getName() + 
+                    "' (created outside of Struts?)");
+            e2.initCause(e);
+            throw e2;
+        }
+        
         Validator validator =
             Resources.initValidator(validationKey, this, application, request,
                 errors, page);